Fixes dataset selector for admin
